[Fix] #3 - 테이블 안만들어지는 이슈 해결
