+++ cgv/src/main/java/org/sopt/server/cgv/controller/RegionController.java
@@ -1,0 +1,4 @@
+package org.sopt.server.cgv.controller;
+
+public class RegionController {
+}
