+++ cgv/src/main/java/org/sopt/server/cgv/repository/MovieRepository.java
@@ -1,0 +1,4 @@
+package org.sopt.server.cgv.repository;
+
+public class MovieRepository {
+}
