+++ cgv/src/main/java/org/sopt/server/cgv/service/MovieService.java
@@ -1,0 +1,4 @@
+package org.sopt.server.cgv.service;
+
+public class MovieService {
+}
