+++ cgv/src/main/java/org/sopt/server/cgv/service/RegionController.java
@@ -1,0 +1,4 @@
+package org.sopt.server.cgv.service;
+
+public class RegionController {
+}
