+++ cgv/src/main/java/org/sopt/server/cgv/service/ScheduleController.java
@@ -1,4 +1,4 @@
-package org.sopt.server.service;
+package org.sopt.server.cgv.service;
 
 public class ScheduleController {
 }
