--- cgv/src/main/java/org/sopt/server/controller/ScheduleController.java
@@ -1,4 +1,0 @@
-package org.sopt.server.controller;
-
-public class ScheduleController {
-}
