--- cgv/src/main/java/org/sopt/server/controller/ScreenController.java
@@ -1,4 +1,0 @@
-package org.sopt.server.controller;
-
-public class ScreenController {
-}
