--- cgv/src/main/java/org/sopt/server/domain/Region.java
@@ -1,4 +1,0 @@
-package org.sopt.server.domain;
-
-public class Region {
-}
