--- cgv/src/main/java/org/sopt/server/repository/RegionRepository.java
@@ -1,4 +1,0 @@
-package org.sopt.server.repository;
-
-public class RegionRepository {
-}
