--- cgv/src/main/java/org/sopt/server/repository/ScreenRepository.java
@@ -1,4 +1,0 @@
-package org.sopt.server.repository;
-
-public class ScreenRepository {
-}
