--- cgv/src/main/java/org/sopt/server/service/MovieService.java
@@ -1,4 +1,0 @@
-package org.sopt.server.service;
-
-public class MovieService {
-}
